requirements.txt
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
decorator==4.0.11
defusedxml==0.5.0
dj-cmd==1.0.0
Django==1.11.2
django-admin-tools==0.8.1
django-allauth==0.32.0
django-application-settings==0.2
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-extensions==1.7.9
django-flat-responsive==1.4.1
django-mptt==0.8.7
django-rest-auth==0.9.1
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
docutils==0.13.1
drfdocs==0.0.11
ecdsa==0.13
fake-factory==9999.9.9
Faker==0.7.18
future==0.16.0
idna==2.5
ipdb==0.10.2
ipython==6.1.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.10.2
Jinja2==2.9.6
MarkupSafe
oauthlib==2.0.2
olefile==0.44
openapi-codec==1.3.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow
prompt-toolkit==1.0.14
psycopg2-binary
psycopg2
ptyprocess==0.5.1
pycrypto==2.6.1
pydot==1.2.3
Pygments==2.2.0
PyJWT==1.5.3
pyparsing==2.2.0
python-crontab==2.2.8
python-dateutil==2.6.1
python-magic==0.4.13
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
rest-social-auth==1.2.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sqlparse==0.2.3
traitlets==4.3.2
typing==3.6.1
uritemplate==3.0.0
urllib3==1.21.1
wcwidth==0.1.7
xlrd==1.1.0